+++ core/sql/regress/privs1/EXPECTED137
@@ -1,8 +1,4 @@
 >>obey TEST137(set_up);
->>grant component privilege "SHOW" on sql_operations to "PUBLIC";
-
---- SQL operation complete.
->>
 >>prepare get_roles from
 +>select substring (auth_db_name,1,20) as role_name
 +>from "_MD_".AUTHS
@@ -1989,9 +1985,6 @@ drop component privilege lib_view_benefactors cascade;
 *** ERROR[1008] Authorization identifier LIB_ROLE_TEST does not exist.
 
 --- SQL operation failed with errors.
->>revoke component privilege "SHOW" on sql_operations from "PUBLIC";
-
---- SQL operation complete.
 >>
 >>revoke role lib_role_test from sql_user5;
