+++ core/sql/sqlcomp/CmpDDLCatErrorCodes.h
@@ -58,7 +58,7 @@ enum CatErrorCode { CAT_FIRST_ERROR = 1000
                   // unused                                       = 1023
                   // unused                                       = 1024
                   , CAT_DEPENDENT_OBJECTS_EXIST                   = 1025
-                  // unused                                       = 1026
+                  , CAT_REG_UNREG_OBJECTS                         = 1026
                   , CAT_COLUMN_PRIVILEGE_NOT_ALLOWED              = 1027
                   , CAT_SCHEMA_IS_NOT_EMPTY                       = 1028
                   , CAT_UNABLE_TO_CREATE_OBJECT                   = 1029
